refactor: 참가자와 딜러를 상속 구조로 변경
